Updates
Conflict indicators indicate if the system has specific conflict types, but this is only for the first conflict in the array.

Systems will display as white if the controlling faction is not tracked. Otherwise, factions have colors associated with them under Game.Manager, and when a new faction is tracked, the viz will update all star systems with the color of the controlling faction.